merge in common development branch

+++ com_biblestudy/admin/lib/assets.php
@@ -19,12 +19,14 @@ defined('_JEXEC') or die;
 class JBSMAssets
 {
 
+	public static $parent_id = null;
+
 	/**
 	 * Fix Assets function.
 	 *
 	 * @return boolean
 	 */
-	public static function fixassets()
+	public static function fixassets ()
 	{
 		$db = JFactory::getDBO();
 
@@ -37,12 +39,7 @@ class JBSMAssets
 		}
 
 		// First get the new parent_id
-		$query = $db->getQuery(true);
-		$query->select('id')
-			->from('#__assets')
-			->where('name = ' . $db->q('com_biblestudy'));
-		$db->setQuery($query);
-		$parent_id = $db->loadResult();
+		self::parentid();
 
 		// Get the names of the JBS tables
 		$objects = self::getassetObjects();
@@ -55,8 +52,8 @@ class JBSMAssets
 			// Get the total number of rows and collect the table into a query
 			$query = $db->getQuery(true);
 			$query->select('j.id, j.asset_id, a.id as aid, a.parent_id')
-				->from($db->qn($object['name']) . ' as j')
-				->leftJoin('#__assets as a ON (a.id = j.asset_id)');
+					->from($db->qn($object['name']) . ' as j')
+					->leftJoin('#__assets as a ON (a.id = j.asset_id)');
 			$db->setQuery($query);
 			$results = $db->loadObjectList();
 
@@ -70,7 +67,7 @@ class JBSMAssets
 				}
 
 				// If there is a jasset_id but no match to the parent_id then a mismatch has occured
-				if ($parent_id != $result->parent_id && $result->asset_id)
+				if (self::$parent_id != $result->parent_id && $result->asset_id)
 				{
 					self::deleteasset($result);
 					self::setasset($result, $object['assetname']);
@@ -82,121 +79,121 @@ class JBSMAssets
 	}
 
 	/**
+	 * Set Parent ID
+	 *
+	 * @return void
+	 */
+	public static function parentid ()
+	{
+		$db = JFactory::getDBO();
+
+		// First get the new parent_id
+		$query = $db->getQuery(true);
+		$query->select('id')
+				->from('#__assets')
+				->where('name = ' . $db->q('com_biblestudy'));
+		$db->setQuery($query);
+		self::$parent_id = $db->loadResult();
+	}
+
+	/**
 	 * Table list Array.
 	 *
 	 * @return array
 	 */
-	protected static function getassetObjects()
+	protected static function getassetObjects ()
 	{
 		$objects = array(
-			array(
-				'name'       => '#__bsms_servers',
-				'titlefield' => 'server_name',
-				'assetname'  => 'server',
-				'realname'   => 'JBS_CMN_SERVERS'
-			),
-			array(
-				'name'       => '#__bsms_folders',
-				'titlefield' => 'foldername',
-				'assetname'  => 'folder',
-				'realname'   => 'JBS_CMN_FOLDERS'
-			),
-			array(
-				'name'       => '#__bsms_studies',
-				'titlefield' => 'studytitle',
-				'assetname'  => 'message',
-				'realname'   => 'JBS_CMN_STUDIES'
-			),
-			array(
-				'name'       => '#__bsms_comments',
-				'titlefield' => 'comment_date',
-				'assetname'  => 'comment',
-				'realname'   => 'JBS_CMN_COMMENTS'
-			),
-			array(
-				'name'       => '#__bsms_locations',
-				'titlefield' => 'location_text',
-				'assetname'  => 'location',
-				'realname'   => 'JBS_CMN_LOCATIONS'
-			),
-			array(
-				'name'       => '#__bsms_media',
-				'titlefield' => 'media_text',
-				'assetname'  => 'mediaimage',
-				'realname'   => 'JBS_CMN_MEDIAIMAGES'
-			),
-			array(
-				'name'       => '#__bsms_mediafiles',
-				'titlefield' => 'filename',
-				'assetname'  => 'mediafile',
-				'realname'   => 'JBS_CMN_MEDIA_FILES'
-			),
-			array(
-				'name'       => '#__bsms_message_type',
-				'titlefield' => 'message_type',
-				'assetname'  => 'messagetype',
-				'realname'   => 'JBS_CMN_MESSAGETYPES'
-			),
-			array(
-				'name'       => '#__bsms_mimetype',
-				'titlefield' => 'mimetext',
-				'assetname'  => 'mimetype',
-				'realname'   => 'JBS_CMN_MIMETYPES'
-			),
-			array(
-				'name'       => '#__bsms_podcast',
-				'titlefield' => 'title',
-				'assetname'  => 'podcast',
-				'realname'   => 'JBS_CMN_PODCASTS'
-			),
-			array(
-				'name'       => '#__bsms_series',
-				'titlefield' => 'series_text',
-				'assetname'  => 'serie',
-				'realname'   => 'JBS_CMN_SERIES'
-			),
-			array(
-				'name'       => '#__bsms_share',
-				'titlefield' => 'name',
-				'assetname'  => 'share',
-				'realname'   => 'JBS_CMN_SOCIAL_NETWORKING_LINKS'
-			),
-			array(
-				'name'       => '#__bsms_teachers',
-				'titlefield' => 'teachername',
-				'assetname'  => 'teacher',
-				'realname'   => 'JBS_CMN_TEACHERS'
-			),
-			array(
-				'name'       => '#__bsms_templates',
-				'titlefield' => 'title',
-				'assetname'  => 'template',
-				'realname'   => 'JBS_CMN_TEMPLATES'
-			),
-			array(
-				'name'       => '#__bsms_topics',
-				'titlefield' => 'topic_text',
-				'assetname'  => 'topic',
-				'realname'   => 'JBS_CMN_TOPICS'
-			),
-			array(
-				'name'       => '#__bsms_styles',
-				'titlefield' => 'filename',
-				'assetname'  => 'style',
-				'realname'   => 'JBS_CMN_STYLES'
-			),
-			array(
-				'name'       => '#__bsms_templatecode',
-				'titlefield' => 'filename',
-				'assetname'  => 'templatecode',
-				'realname'   => 'JBS_CMN_TEMPLATECODE'
-			),
-			array(
-				'name'       => '#__bsms_admin',
-				'titlefield' => 'id',
-				'assetname'  => 'admin',
-				'realname'   => 'JBS_CMN_ADMINISTRATION'
-			)
+				array(
+						'name'       => '#__bsms_servers',
+						'titlefield' => 'server_name',
+						'assetname'  => 'server',
+						'realname'   => 'JBS_CMN_SERVERS'
+				),
+				array(
+						'name'       => '#__bsms_studies',
+						'titlefield' => 'studytitle',
+						'assetname'  => 'message',
+						'realname'   => 'JBS_CMN_STUDIES'
+				),
+				array(
+						'name'       => '#__bsms_comments',
+						'titlefield' => 'comment_date',
+						'assetname'  => 'comment',
+						'realname'   => 'JBS_CMN_COMMENTS'
+				),
+				array(
+						'name'       => '#__bsms_locations',
+						'titlefield' => 'location_text',
+						'assetname'  => 'location',
+						'realname'   => 'JBS_CMN_LOCATIONS'
+				),
+				array(
+						'name'       => '#__bsms_mediafiles',
+						'titlefield' => 'filename',
+						'assetname'  => 'mediafile',
+						'realname'   => 'JBS_CMN_MEDIA_FILES'
+				),
+				array(
+						'name'       => '#__bsms_message_type',
+						'titlefield' => 'message_type',
+						'assetname'  => 'messagetype',
+						'realname'   => 'JBS_CMN_MESSAGETYPES'
+				),
+				array(
+						'name'       => '#__bsms_podcast',
+						'titlefield' => 'title',
+						'assetname'  => 'podcast',
+						'realname'   => 'JBS_CMN_PODCASTS'
+				),
+				array(
+						'name'       => '#__bsms_series',
+						'titlefield' => 'series_text',
+						'assetname'  => 'serie',
+						'realname'   => 'JBS_CMN_SERIES'
+				),
+				array(
+						'name'       => '#__bsms_share',
+						'titlefield' => 'name',
+						'assetname'  => 'share',
+						'realname'   => 'JBS_CMN_SOCIAL_NETWORKING_LINKS'
+				),
+				array(
+						'name'       => '#__bsms_teachers',
+						'titlefield' => 'teachername',
+						'assetname'  => 'teacher',
+						'realname'   => 'JBS_CMN_TEACHERS'
+				),
+				array(
+						'name'       => '#__bsms_templates',
+						'titlefield' => 'title',
+						'assetname'  => 'template',
+						'realname'   => 'JBS_CMN_TEMPLATES'
+				),
+				array(
+						'name'       => '#__bsms_topics',
+						'titlefield' => 'topic_text',
+						'assetname'  => 'topic',
+						'realname'   => 'JBS_CMN_TOPICS'
+				),
+				array(
+						'name'       => '#__bsms_styles',
+						'titlefield' => 'filename',
+						'assetname'  => 'style',
+						'realname'   => 'JBS_CMN_STYLES'
+				),
+				array(
+						'name'       => '#__bsms_templatecode',
+						'titlefield' => 'filename',
+						'assetname'  => 'templatecode',
+						'realname'   => 'JBS_CMN_TEMPLATECODE'
+				),
+				array(
+						'name'       => '#__bsms_admin',
+						'titlefield' => 'id',
+						'assetname'  => 'admin',
+						'realname'   => 'JBS_CMN_ADMINISTRATION'
+				)
 		);
 
 		return $objects;
@@ -205,22 +202,30 @@ class JBSMAssets
 	/**
 	 * Set Asset
 	 *
-	 * @param   object $data       Data
-	 * @param   string $assetname  Asset Name
+	 * @param   object  $data       Data
+	 * @param   string  $assetname  Asset Name
 	 *
 	 * @return boolean
 	 */
-	private static function setasset($data, $assetname)
+	private static function setasset ($data, $assetname)
 	{
-		$db = JFactory::getDBO();
 		JTable::addIncludePath(JPATH_COMPONENT_ADMINISTRATOR . '/tables');
-		$table = JTable::getInstance($assetname, 'Table', array('dbo' => $db));
+		$table = JTable::getInstance($assetname, 'Table');
 
 		if ($data->id)
 		{
 			try
 			{
-				$table->load($data->id);
+				if ($assetname == 'mediafile')
+				{
+					$columns = array('media_image', 'special', 'filename', 'size', 'mime_type', 'mediacode', 'link_type',
+							'docMan_id', 'article_id', 'virtueMart_id', 'player', 'popup', 'server', 'internal_viewer', 'path');
+					foreach ($columns as $col)
+					{
+						unset($table->$col);
+					}
+				}
+				$table->load($data->id, false);
 			}
 			catch (Exception $e)
 			{
@@ -228,12 +233,7 @@ class JBSMAssets
 
 				return false;
 			}
-			if (@!$table->store())
-			{
-				JFactory::getApplication()->enqueueMessage(JText::sprintf('JBS_INS_SQL_UPDATE_ERRORS', $db->stderr(true)), 'error');
-
-				return false;
-			}
+			$table->store();
 		}
 
 		return true;
@@ -242,29 +242,23 @@ class JBSMAssets
 	/**
 	 * Delete assets
 	 *
-	 * @param   object $data  Data
+	 * @param   object  $data  Data
 	 *
 	 * @return boolean
 	 */
-	private static function deleteasset($data)
+	private static function deleteasset ($data)
 	{
 		$db = JFactory::getDBO();
 
 		if (isset($data->jasset_id))
 		{
-			if ($data->jasset_id >= 2)
+			if ($data->jasset_id >= 2 && $data->jasset_id != self::$parent_id)
 			{
 				$query = $db->getQuery(true);
 				$query->delete('#__assets')
-					->where('id = ' . $db->quote($data->jasset_id));
+						->where('id = ' . $db->quote($data->jasset_id));
 				$db->setQuery($query);
-
-				if (!$db->execute())
-				{
-					JFactory::getApplication()->enqueueMessage(JText::sprintf('JBS_INS_SQL_UPDATE_ERRORS', $db->stderr(true)), 'error');
-
-					return false;
-				}
+				$db->execute();
 			}
 
 			return true;
@@ -280,19 +274,17 @@ class JBSMAssets
 	 *
 	 * @return array
 	 */
-	public static function checkAssets()
+	public static function checkAssets ()
 	{
 		$return = array();
 		$db     = JFactory::getDBO();
 		$result = new stdClass;
 
 		// First get the new parent_id
-		$query = $db->getQuery(true);
-		$query->select('id')
-			->from('#__assets')
-			->where('name = ' . $db->q('com_biblestudy'));
-		$db->setQuery($query);
-		$parent_id = $db->loadResult();
+		if (!self::$parent_id)
+		{
+			self::parentid();
+		}
 
 		// Get the names of the JBS tables
 		$objects = self::getassetObjects();
@@ -304,8 +296,8 @@ class JBSMAssets
 			// Get the total number of rows and collect the table into a query
 			$query = $db->getQuery(true);
 			$query->select('j.id as jid, j.asset_id as jasset_id, a.id as aid, a.parent_id')
-				->from($db->qn($object['name']) . ' as j')
-				->leftJoin('#__assets as a ON (a.id = j.asset_id)');
+					->from($db->qn($object['name']) . ' as j')
+					->leftJoin('#__assets as a ON (a.id = j.asset_id)');
 			$db->setQuery($query);
 			$results     = $db->loadObjectList();
 			$nullrows    = 0;
@@ -322,26 +314,26 @@ class JBSMAssets
 					$nullrows++;
 				}
 				// If there is a jasset_id but no match to the parent_id then a mismatch has occured
-				if ($parent_id != $result->parent_id && $result->jasset_id)
+				if (self::$parent_id != $result->parent_id && $result->jasset_id)
 				{
 					$nomatchrows++;
 				}
 				// If $parent_id and $result->parent_id match then everything is okay
-				if ($parent_id == $result->parent_id)
+				if (self::$parent_id == $result->parent_id)
 				{
 					$matchrows++;
 				}
 			}
 			$return[] = array(
-				'realname'         => $object['realname'],
-				'numrows'          => $numrows,
-				'nullrows'         => $nullrows,
-				'matchrows'        => $matchrows,
-				'nomatchrows'      => $nomatchrows,
-				'parent_id'        => $parent_id,
-				'result_parent_id' => $result->parent_id,
-				'id'               => $result->jid,
-				'assetid'          => $result->jasset_id
+					'realname'         => $object['realname'],
+					'numrows'          => $numrows,
+					'nullrows'         => $nullrows,
+					'matchrows'        => $matchrows,
+					'nomatchrows'      => $nomatchrows,
+					'parent_id'        => self::$parent_id,
+					'result_parent_id' => $result->parent_id,
+					'id'               => $result->jid,
+					'assetid'          => $result->jasset_id
 			);
 		}
